Move database connection into db_helpers
