+++ ingestion_server/ingestion_server/ingest.py
@@ -26,7 +26,7 @@ from psycopg2.sql import SQL, Identifier, Literal
 from ingestion_server import slack
 from ingestion_server.cleanup import clean_image_data
 from ingestion_server.constants.internal_types import ApproachType
-from ingestion_server.indexer import database_connect
+from ingestion_server.db_helpers import database_connect
 from ingestion_server.queries import (
     get_copy_data_query,
     get_create_ext_query,
